backend/requirements.txt
# Django Core
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0

# Database
psycopg2-binary>=2.9.11
psycopg>=3.2.12
psycopg-binary>=3.2.12
psycopg-pool>=3.2.7

# Redis
redis>=7.0.0

# API Documentation
drf-yasg>=1.21.7

# Celery for async tasks
celery>=5.3.0
django-celery-beat>=2.5.0

# Environment variables
python-dotenv>=1.0.0

# HTTP requests
requests>=2.32.5
httpx>=0.24.0  # For async HTTP (future use)

# Security
pycryptodome>=3.23.0

# Utilities
typing-extensions>=4.15.0

# Development
django-debug-toolbar>=4.2.0  # Only for dev

